# Use l1-Jacobi smoother for ADS with CUDA [ads-cuda-use-jacobi]